+++ resources/views/components/multiSelect.blade.php
@@ -1,1 +1,5 @@
-<multiselect :values="{{json_encode($values)}}"></multiselect>
+@isset($preSelect)
+    <multiselect :values="{{json_encode($values)}}" :pre-select="{{json_encode($preSelect)}}"></multiselect>
+@else
+    <multiselect :values="{{json_encode($values)}}"></multiselect>
+@endif