+++ src/View/Components/MultiSelect.php
@@ -7,17 +7,12 @@ use Illuminate\View\Component;
 
 class MultiSelect extends Component {
 
-    public $values;
-
     /**
      * Create a new component instance.
      *
      * @return void
      */
-    public function __construct($values) {
-//        $this->values = $values;
-        $this->values = ['test1', 'test2'];
-    }
+    public function __construct() { }
 
     /**
      * Get the view / contents that represent the component.